Move all Access-related stuff to UserAccess and delegate from User
