+++ app/views/email_authentications/invitations/new.html.erb
@@ -6,7 +6,7 @@
         <%= form.text_field :full_name, id: :full_name, required: true, autocomplete: "off", label: "Full Name *" %>
         <%= form.text_field :email, id: :email, required: true, label: "Email *" %>
         <%= form.text_field :role, id: :role, required: true, help: "CVHO, STS, State Official etc.", label: 'Job Title *' %>
-        <%= form.select :access_levels, current_admin.grantable_access_levels.map { |_, data| data.values_at(:name, :id) }, id: :access_levels, required: true, label: "Access *" %>
+        <%= form.select :access_levels, current_admin.permitted_access_levels.map { |_, data| data.values_at(:name, :id) }, id: :access_levels, required: true, label: "Access *" %>
 
         <%= form.label "Facility Access *" %>
         <div id="facility-access">
